Description
 Definition:
 Insertion sort is a simple sorting algorithm that works similar to the way you sort playing cards in your hands. The array is virtually split into a sorted and an unsorted part. Values from the unsorted part are picked and placed at the correct position in the sorted part.
 
Explanation:
Here a class InsertionSort is defined. Inside the class a main method InsertionSort is declared. A variable, size is defined as an integer which is assigned to length of the array inside the main method. A for loop is constructed for insertion sort, inside the for loop a variable step is assigned as 1 initially which is of type integer, comparing it with size of the array which it should be less than total size. A variable key is initialized which is of type integer and is assigned to array[step] and a variable j of type integer is initialized which is assigned as step-1.  
Now a while loop is declared to compare key with each element on the left of it until an element smaller than it is found. 
For descending order, change key<array[j] to key>array[j].
After the while loop has completed its task place key at after the element just smaller than it.
Then inside the void main function an array 'data' of size 4 is initialized of type integer with array elements 9,5,1,4,3. An object is created of class InsertionSort and the object is calling the main method for sorted array elements using data i.e., insertionSort(data).
Finally, the sorted array elements are executed by giving argument ("Sorted Array in Ascending Order: ").